Add optional "comment" argument to the generator method.
If comment is specified and not None, then a JPEG "COM" field is added
to the generated JPEG, containing the specified comment. It's possible
that JPEGs with comments might be attractive to abusive web crawlers,
and it's trivially easy to add the COM field.

Option suggested by "coldclimate" on the Fediverse.